LayoutTests/html5lib/run-test18-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
resources/tests18.dat: PASS
